Check oldusernames before  404 redirect on views
issue #706

+++ follow/views.py
@@ -20,7 +20,7 @@
 #     See AUTHORS file.
 #
 
-from django.http import HttpResponse
+from django.http import HttpResponse, Http404
 from django.shortcuts import get_object_or_404, render
 from django.contrib.auth.decorators import login_required
 from follow import follow_utils
@@ -30,10 +30,13 @@ from django.contrib.auth.models import User
 from datetime import datetime, timedelta
 from collections import OrderedDict
 from socket import error as socket_error
+from utils.username import get_user_from_oldusername
 
 
 def following_users(request, username):
-    user = get_object_or_404(User, username=username)
+    user = get_user_from_oldusername(username)
+    if user == None:
+        raise Http404
     is_owner = False
     if request.user.is_authenticated:
         is_owner = request.user == user
@@ -46,7 +49,9 @@ def following_users(request, username):
 
 
 def followers(request, username):
-    user = get_object_or_404(User, username=username)
+    user = get_user_from_oldusername(username)
+    if user == None:
+        raise Http404
     is_owner = False
     if request.user.is_authenticated:
         is_owner = request.user == user
@@ -59,7 +64,9 @@ def followers(request, username):
 
 
 def following_tags(request, username):
-    user = get_object_or_404(User, username=username)
+    user = get_user_from_oldusername(username)
+    if user == None:
+        raise Http404
     is_owner = False
     if request.user.is_authenticated:
         is_owner = request.user == user
@@ -86,7 +93,9 @@ def following_tags(request, username):
 def follow_user(request, username):
     # create following user item relation
     user_from = request.user
-    user_to = get_object_or_404(User, username=username)
+    user_to = get_user_from_oldusername(username)
+    if user_to == None:
+        raise Http404
     FollowingUserItem.objects.get_or_create(user_from=user_from, user_to=user_to)
     return HttpResponse()
 
@@ -94,7 +103,9 @@ def follow_user(request, username):
 @login_required
 def unfollow_user(request, username):
     user_from = request.user
-    user_to = get_object_or_404(User, username=username)
+    user_to = get_user_from_oldusername(username)
+    if user_to == None:
+        raise Http404
     try:
         FollowingUserItem.objects.get(user_from=user_from, user_to=user_to).delete()
     except FollowingUserItem.DoesNotExist:
